removed all errors
