+++ ClassLibrary1/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace ClassLibrary1
+{
+    public class Class1
+    {
+    }
+}
