+++ kmd-logic-consumption-client.sln
@@ -9,11 +9,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{BB2479FF-5DE
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sample", "sample", "{9D933673-157D-4C45-8C89-81085A5E99F4}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kmd.Logic.Consumption.Client.Sample", "sample\kmd-logic-consumption-client.sample\Kmd.Logic.Consumption.Client.Sample.csproj", "{33966829-2B7F-4F84-8D23-81EBCB180D6B}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{9B0D548B-CCEB-4C61-B5D3-C466534BF4BD}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kmd.Logic.Consumption.Client.AuditClient", "src\Kmd.Logic.Consumption.Client.AuditClient\Kmd.Logic.Consumption.Client.AuditClient.csproj", "{62EBC8E9-233E-49E9-A27A-5C91554B0C08}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kmd.Logic.Consumption.Client.Sample", "sample\Kmd.Logic.Consumption.Client.Sample\Kmd.Logic.Consumption.Client.Sample.csproj", "{E0ECBFFE-8526-4E0B-B02B-7561AC14E8EB}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kmd.Logic.Consumption.Client.Tests", "test\Kmd.Logic.Consumption.Client.Tests\Kmd.Logic.Consumption.Client.Tests.csproj", "{42BABB5D-735A-4A05-B5D5-351A32C1B6C3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,22 +27,27 @@ Global
 		{82C1F757-109A-40A7-8268-84CC3A249DFE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{82C1F757-109A-40A7-8268-84CC3A249DFE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{82C1F757-109A-40A7-8268-84CC3A249DFE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{33966829-2B7F-4F84-8D23-81EBCB180D6B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{33966829-2B7F-4F84-8D23-81EBCB180D6B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{33966829-2B7F-4F84-8D23-81EBCB180D6B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{33966829-2B7F-4F84-8D23-81EBCB180D6B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{62EBC8E9-233E-49E9-A27A-5C91554B0C08}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{62EBC8E9-233E-49E9-A27A-5C91554B0C08}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{62EBC8E9-233E-49E9-A27A-5C91554B0C08}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{62EBC8E9-233E-49E9-A27A-5C91554B0C08}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E0ECBFFE-8526-4E0B-B02B-7561AC14E8EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E0ECBFFE-8526-4E0B-B02B-7561AC14E8EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E0ECBFFE-8526-4E0B-B02B-7561AC14E8EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E0ECBFFE-8526-4E0B-B02B-7561AC14E8EB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{42BABB5D-735A-4A05-B5D5-351A32C1B6C3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{42BABB5D-735A-4A05-B5D5-351A32C1B6C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{42BABB5D-735A-4A05-B5D5-351A32C1B6C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{42BABB5D-735A-4A05-B5D5-351A32C1B6C3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{82C1F757-109A-40A7-8268-84CC3A249DFE} = {BB2479FF-5DEC-4A78-A59F-C917F60F377B}
-		{33966829-2B7F-4F84-8D23-81EBCB180D6B} = {9D933673-157D-4C45-8C89-81085A5E99F4}
 		{62EBC8E9-233E-49E9-A27A-5C91554B0C08} = {BB2479FF-5DEC-4A78-A59F-C917F60F377B}
+		{E0ECBFFE-8526-4E0B-B02B-7561AC14E8EB} = {9D933673-157D-4C45-8C89-81085A5E99F4}
+		{42BABB5D-735A-4A05-B5D5-351A32C1B6C3} = {9B0D548B-CCEB-4C61-B5D3-C466534BF4BD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C03872C2-92F5-48B8-AA12-76BEFEEA0319}
