+++ sample/Kmd.Logic.Consumption.Client.Sample/Kmd.Logic.Consumption.Client.Sample.csproj
@@ -1,10 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AssemblyName>Kmd.Logic.Consumption.Client.Sample</AssemblyName>
-    <RootNamespace>Kmd.Logic.Consumption.Client.Sample</RootNamespace>
+    <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
